SEO, Hover Effects & Admin Actions
SEO:
Added better SEO using alts on images. and fixing some other text.

Hover Effects:
When hovering over an input have you a nice indicator you are hovering over. If you are editing it then it will also have a different indicator. If you cant edit it, it will be darkend.

Admin Actions:
Added the abilty for admins to edit the pages on the site (currently just the DB info) and also see all the pages ono the site in a nice table.